requirements.txt
crewai==0.130.0
crewai-tools>=0.47.1
fastapi==0.110.3
uvicorn>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
langchain-community>=0.2.0
pydantic>=2.4.2
pydantic_core>=2.8.0
openai>=1.30.5
numpy>=1.26.4
pandas>=2.2.2
pillow>=10.3.0
pypdf>=3.0.0
